Refactor dora worker metastore and remove on-heap cache
1. Support more rocksdb options in dora meta store
2. Add a microbench to profile the dora meta store result
3. Removed the memory metadata cache based on the profiling result  
4. Introduced a DoraMetaManager to handle data loading/ page cache invalidation
5. Introduced fingerprint to determine if file contents are changed.  

Performance testing: 
Microbench result on a single worker with a metastore storing 10m metadata entries
Rocks + 32MB rocks block cache:
Read op/s 81k (1 thread)/ 143k (16 threads)
Write op/s 285k (1 thread)/ 196k (16 threads)

Rocks + 1GB rocks block cache:
Read op/s 163k (1 thread)/ 480k (16 threads)
Write op/s 286k (1 thread)/ 134k (16 threads)

Rocks + 4GB on-heap cache:
Read ops 560k (1 thread)/ OOM (16 threads)
Write ops 330k (1 thread)/ 363k (16 threads)
-> Observed severe GC issue when using an on-heap cache 
<img width="1386" alt="Screen Shot 2023-05-29 at 4 16 52 PM" src="https://github.com/Alluxio/alluxio/assets/6771554/b156d4c9-97ba-4216-929c-4c81b4f5ce3c">


Both rocksDB built-in cache approaches meet our performance requirement and given that the GC issue of on-heap cache, we decided to abandon the on-heap memory cache. 

			pr-link: Alluxio/alluxio#17458

+++ dora/core/server/common/src/main/java/alluxio/rocks/RocksStore.java
@@ -339,7 +339,7 @@ public final class RocksStore implements Closeable {
     if (Configuration.isSet(cacheSize)) {
       shoudSetConfig = true;
       // Set the inodes column options
-      Cache inodeCache = new LRUCache(Configuration.getInt(cacheSize));
+      Cache inodeCache = new LRUCache(Configuration.getLong(cacheSize));
       toClose.add(inodeCache);
       blockConfig.setBlockCache(inodeCache);
     }
